Refactoring post processing pipeline
